+++ extensions/easydict/src/consts.ts
@@ -2,44 +2,15 @@
  * @author: tisfeng
  * @createTime: 2022-06-24 22:36
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-01 11:05
+ * @lastEditTime: 2022-07-16 16:18
  * @fileName: consts.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
-import { LanguageDetectType } from "./detectLanguage";
 import { LanguageItem, RequestErrorInfo } from "./types";
 
 export const clipboardQueryTextKey = "clipboardQueryTextKey";
-
-/**
- *
- */
-export const maxInputTextLength = 2000;
-
-export enum SectionType {
-  Translation = "Translate",
-  Explanations = "Explanation",
-  Forms = "Forms and Tenses",
-  WebTranslation = "Web Translation",
-  WebPhrase = "Web Phrase",
-}
-
-export enum TranslateType {
-  Youdao = "Youdao",
-  Baidu = "Baidu",
-  Tencent = "Tencent",
-  Caiyun = "Caiyun",
-  Apple = "Apple",
-}
-
-export enum DicionaryType {
-  Youdao = "Youdao Dictionary",
-  Iciba = "Iciba Dictionary",
-}
-
-export type RequestType = TranslateType | DicionaryType | LanguageDetectType;
 
 export enum YoudaoRequestStateCode {
   Success = "0",
@@ -94,7 +65,10 @@ export function getYoudaoErrorInfo(errorCode: string): RequestErrorInfo {
 }
 
 /**
- * 语言代码列表
+ * Language item list.
+ * Currently support 22 languages, includes two Chinese languages.
+ * TODO: add more translation supported languages.
+ * TODO: add more apple detect supported languages.
  */
 export const languageItemList: LanguageItem[] = [
   {
@@ -112,6 +86,7 @@ export const languageItemList: LanguageItem[] = [
     youdaoLanguageId: "zh-CHS",
     appleLanguageId: "zh_CN",
     appleChineseLanguageTitle: "中文",
+    deepLSourceLanguageId: "ZH",
     francLanguageId: "cmn",
     aliyunLanguageId: "zh",
     tencentLanguageId: "zh",
@@ -129,7 +104,6 @@ export const languageItemList: LanguageItem[] = [
     aliyunLanguageId: "zh-tw",
     tencentLanguageId: "zh-TW",
     baiduLanguageId: "cht",
-    caiyunLanguageId: "zh",
     googleLanguageId: "zh-TW",
     languageTitle: "Chinese-Traditional",
     voiceList: ["Ting-Ting"],
@@ -138,6 +112,8 @@ export const languageItemList: LanguageItem[] = [
     youdaoLanguageId: "en",
     appleLanguageId: "en_US",
     appleChineseLanguageTitle: "英语",
+    deepLSourceLanguageId: "EN",
+    deepLTargetLanguageId: "EN-US", // "EN-GB" "EN-US"
     francLanguageId: "eng",
     aliyunLanguageId: "en",
     tencentLanguageId: "en",
@@ -153,6 +129,7 @@ export const languageItemList: LanguageItem[] = [
     youdaoLanguageId: "ja",
     appleLanguageId: "ja_JP",
     appleChineseLanguageTitle: "日语",
+    deepLSourceLanguageId: "JA",
     francLanguageId: "jpn",
     aliyunLanguageId: "ja",
     tencentDetectLanguageId: "jp",
@@ -180,6 +157,7 @@ export const languageItemList: LanguageItem[] = [
     youdaoLanguageId: "fr",
     appleLanguageId: "fr_FR",
     appleChineseLanguageTitle: "法语",
+    deepLSourceLanguageId: "FR",
     francLanguageId: "fra",
     aliyunLanguageId: "fr",
     tencentLanguageId: "fr",
@@ -193,6 +171,7 @@ export const languageItemList: LanguageItem[] = [
     youdaoLanguageId: "es",
     appleLanguageId: "es_ES",
     appleChineseLanguageTitle: "西班牙语",
+    deepLSourceLanguageId: "ES",
     francLanguageId: "spa",
     aliyunLanguageId: "es",
     tencentLanguageId: "es",
@@ -202,9 +181,23 @@ export const languageItemList: LanguageItem[] = [
     voiceList: ["Jorge", "Juan", "Diego", "Monica", "Paulina"],
   },
   {
+    youdaoLanguageId: "pt",
+    appleLanguageId: "pt_BR",
+    appleChineseLanguageTitle: "葡萄牙语",
+    deepLSourceLanguageId: "PT",
+    deepLTargetLanguageId: "PT-PT", // "PT-PT" "PT-BR"
+    francLanguageId: "por",
+    aliyunLanguageId: "pt",
+    tencentLanguageId: "pt",
+    baiduLanguageId: "pt",
+    languageTitle: "Portuguese",
+    voiceList: ["Joana", "Luciana"],
+  },
+  {
     youdaoLanguageId: "it",
     appleLanguageId: "it_IT",
     appleChineseLanguageTitle: "意大利语",
+    deepLSourceLanguageId: "IT",
     francLanguageId: "ita",
     aliyunLanguageId: "it",
     tencentLanguageId: "it",
@@ -216,6 +209,7 @@ export const languageItemList: LanguageItem[] = [
     youdaoLanguageId: "de",
     appleLanguageId: "de_DE",
     appleChineseLanguageTitle: "德语",
+    deepLSourceLanguageId: "DE",
     francLanguageId: "deu",
     aliyunLanguageId: "de",
     tencentLanguageId: "de",
@@ -225,20 +219,10 @@ export const languageItemList: LanguageItem[] = [
     voiceList: ["Anna"],
   },
   {
-    youdaoLanguageId: "pt",
-    appleLanguageId: "pt_BR",
-    appleChineseLanguageTitle: "葡萄牙语",
-    francLanguageId: "por",
-    aliyunLanguageId: "pt",
-    tencentLanguageId: "pt",
-    baiduLanguageId: "pt",
-    languageTitle: "Portuguese",
-    voiceList: ["Joana", "Luciana"],
-  },
-  {
     youdaoLanguageId: "ru",
     appleLanguageId: "ru_RU",
     appleChineseLanguageTitle: "俄语",
+    deepLSourceLanguageId: "RU",
     francLanguageId: "rus",
     aliyunLanguageId: "ru",
     tencentLanguageId: "ru",
@@ -258,6 +242,24 @@ export const languageItemList: LanguageItem[] = [
     voiceList: ["Maged"],
   },
   {
+    youdaoLanguageId: "sv",
+    deepLSourceLanguageId: "SV",
+    francLanguageId: "swe",
+    aliyunLanguageId: "sv",
+    baiduLanguageId: "swe",
+    languageTitle: "Swedish",
+    voiceList: ["Alva"],
+  },
+  {
+    youdaoLanguageId: "ro",
+    deepLSourceLanguageId: "RO",
+    francLanguageId: "ron",
+    aliyunLanguageId: "ro",
+    baiduLanguageId: "rom",
+    languageTitle: "Romanian",
+    voiceList: ["Ioana"],
+  },
+  {
     youdaoLanguageId: "th",
     francLanguageId: "tha",
     aliyunLanguageId: "th",
@@ -267,31 +269,8 @@ export const languageItemList: LanguageItem[] = [
     voiceList: ["Kanya"],
   },
   {
-    youdaoLanguageId: "sv",
-    francLanguageId: "swe",
-    aliyunLanguageId: "sv",
-    baiduLanguageId: "swe",
-    languageTitle: "Swedish",
-    voiceList: ["Alva"],
-  },
-  {
-    youdaoLanguageId: "nl",
-    francLanguageId: "nld",
-    aliyunLanguageId: "nl",
-    baiduLanguageId: "nl",
-    languageTitle: "Dutch",
-    voiceList: ["Ellen", "Xander"],
-  },
-  {
-    youdaoLanguageId: "ro",
-    francLanguageId: "ron",
-    aliyunLanguageId: "ro",
-    baiduLanguageId: "rom",
-    languageTitle: "Romanian",
-    voiceList: ["Ioana"],
-  },
-  {
     youdaoLanguageId: "sk",
+    deepLSourceLanguageId: "SK",
     francLanguageId: "slk",
     aliyunLanguageId: "sk",
     baiduLanguageId: "slo",
@@ -299,7 +278,17 @@ export const languageItemList: LanguageItem[] = [
     voiceList: ["Laura"],
   },
   {
+    youdaoLanguageId: "nl",
+    deepLSourceLanguageId: "NL",
+    francLanguageId: "nld",
+    aliyunLanguageId: "nl",
+    baiduLanguageId: "nl",
+    languageTitle: "Dutch",
+    voiceList: ["Ellen", "Xander"],
+  },
+  {
     youdaoLanguageId: "hu",
+    deepLSourceLanguageId: "HU",
     francLanguageId: "hun",
     aliyunLanguageId: "hu",
     baiduLanguageId: "hu",
@@ -308,6 +297,7 @@ export const languageItemList: LanguageItem[] = [
   },
   {
     youdaoLanguageId: "el",
+    deepLSourceLanguageId: "EL",
     francLanguageId: "ell",
     aliyunLanguageId: "el",
     baiduLanguageId: "el",
@@ -316,6 +306,7 @@ export const languageItemList: LanguageItem[] = [
   },
   {
     youdaoLanguageId: "da",
+    deepLSourceLanguageId: "DA",
     francLanguageId: "dan",
     aliyunLanguageId: "da",
     baiduLanguageId: "dan",
@@ -324,6 +315,7 @@ export const languageItemList: LanguageItem[] = [
   },
   {
     youdaoLanguageId: "fi",
+    deepLSourceLanguageId: "FI",
     francLanguageId: "fin",
     aliyunLanguageId: "fi",
     baiduLanguageId: "fin",
@@ -332,6 +324,7 @@ export const languageItemList: LanguageItem[] = [
   },
   {
     youdaoLanguageId: "pl",
+    deepLSourceLanguageId: "PL",
     francLanguageId: "pol",
     aliyunLanguageId: "pl",
     baiduLanguageId: "pl",
@@ -340,6 +333,7 @@ export const languageItemList: LanguageItem[] = [
   },
   {
     youdaoLanguageId: "cs",
+    deepLSourceLanguageId: "CS",
     francLanguageId: "ces",
     aliyunLanguageId: "cs",
     baiduLanguageId: "cs",