+++ extensions/easydict/src/utils.ts
@@ -2,17 +2,16 @@
  * @author: tisfeng
  * @createTime: 2022-06-26 11:13
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-03 17:08
+ * @lastEditTime: 2022-07-15 22:07
  * @fileName: utils.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
-import { Clipboard, environment, getApplications, getPreferenceValues, LocalStorage } from "@raycast/api";
+import { Clipboard, getApplications, getPreferenceValues, LocalStorage } from "@raycast/api";
 import { eudicBundleId } from "./components";
 import { clipboardQueryTextKey, languageItemList } from "./consts";
 import { LanguageItem, MyPreferences, QueryRecoredItem, QueryWordInfo, TranslateFormatResult } from "./types";
-import CryptoJS from "crypto-js";
 
 // Time interval for automatic query of the same clipboard text, avoid frequently querying the same word. Default 10min
 export const clipboardQueryInterval = 10 * 60 * 1000;
@@ -24,27 +23,6 @@ export const myPreferences: MyPreferences = getPreferenceValues();
 export const defaultLanguage1 = getLanguageItemFromYoudaoId(myPreferences.language1) as LanguageItem;
 export const defaultLanguage2 = getLanguageItemFromYoudaoId(myPreferences.language2) as LanguageItem;
 export const preferredLanguages = [defaultLanguage1, defaultLanguage2];
-
-const defaultEncrytedYoudaoAppId = "U2FsdGVkX19SpBCGxMeYKP0iS1PWKmvPeqIYNaZjAZC142Y5pLrOskw0gqHGpVS1";
-const defaultEncrytedYoudaoAppKey =
-  "U2FsdGVkX1/JF2ZMngmTw8Vm+P0pHWmHKLQhGpUtYiDc0kLZl6FKw1Vn3hMyl7iL7owwReGJCLsovDxztZKb9g==";
-export const defaultYoudaoAppId = myDecrypt(defaultEncrytedYoudaoAppId);
-export const defaultYoudaoAppSecret = myDecrypt(defaultEncrytedYoudaoAppKey);
-
-const defaultEncryptedBaiduAppId = "U2FsdGVkX1/QHkSw+8qxr99vLkSasBfBRmA6Kb5nMyjP8IJazM9DcOpd3cOY6/il";
-const defaultEncryptedBaiduAppSecret = "U2FsdGVkX1+a2LbZ0+jntJTQjpPKUNWGrlr4NSBOwmlah7iP+w2gefq1UpCan39J";
-export const defaultBaiduAppId = myDecrypt(defaultEncryptedBaiduAppId);
-export const defaultBaiduAppSecret = myDecrypt(defaultEncryptedBaiduAppSecret);
-
-const defaultEncryptedTencentSecretId =
-  "U2FsdGVkX19lHBVXE+CEZI9cENSToLIGzHDsUIE+RyvIC66rgxumDmpYPDY4MdaTSbrq7MIyDvtgXaLvzijYSg==";
-const defaultEncryptedTencentSecretKey =
-  "U2FsdGVkX1+N6wDYXNiUISwKOM97cY03RjXmC+0+iodFo3b4NTNC1J8RR6xqcbdyF7z3Z2yQRMHHxn4m02aUvA==";
-export const defaultTencentSecretId = myDecrypt(defaultEncryptedTencentSecretId);
-export const defaultTencentSecretKey = myDecrypt(defaultEncryptedTencentSecretKey);
-
-const defaultEncryptedCaiyunToken = "U2FsdGVkX1+ihWvHkAfPMrWHju5Kg4EXAm1AVbXazEeHaXE1jdeUzZZrhjdKmS6u";
-export const defaultCaiyunToken = myDecrypt(defaultEncryptedCaiyunToken);
 
 export function getLanguageItemFromYoudaoId(youdaoLanguageId: string): LanguageItem {
   for (const langItem of languageItemList) {
@@ -76,6 +54,18 @@ export function getLanguageItemFromTencentId(tencentLanguageId: string): Languag
 export function getLanguageItemFromAppleChineseTitle(chineseTitle: string): LanguageItem {
   for (const langItem of languageItemList) {
     if (langItem.appleChineseLanguageTitle === chineseTitle) {
+      return langItem;
+    }
+  }
+  return languageItemList[0];
+}
+
+/**
+ * Return language item from deepL language id, if not found, return auto language item
+ */
+export function getLanguageItemFromDeepLSourceId(deepLLanguageId: string): LanguageItem {
+  for (const langItem of languageItemList) {
+    if (langItem.deepLSourceLanguageId === deepLLanguageId) {
       return langItem;
     }
   }
@@ -262,21 +252,6 @@ export function checkIsInstalledEudic(setIsInstalledEudic: (isInstalled: boolean
       traverseAllInstalledApplications(setIsInstalledEudic);
     }
   });
-}
-
-export function myEncrypt(text: string) {
-  // console.warn("encrypt:", text);
-  const ciphertext = CryptoJS.AES.encrypt(text, environment.extensionName).toString();
-  // console.warn("ciphertext: ", ciphertext);
-  return ciphertext;
-}
-
-export function myDecrypt(ciphertext: string) {
-  // console.warn("decrypt:", ciphertext);
-  const bytes = CryptoJS.AES.decrypt(ciphertext, environment.extensionName);
-  const originalText = bytes.toString(CryptoJS.enc.Utf8);
-  // console.warn("originalText: ", originalText);
-  return originalText;
 }
 
 export function isShowMultipleTranslations(formatResult: TranslateFormatResult) {